python-service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sentence-transformers==2.2.2
transformers==4.35.0
faiss-cpu==1.7.4
python-multipart==0.0.6
pydantic==1.10.7
requests==2.31.0
